@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -19,7 +14,7 @@ namespace PopForums.Sample
 {
 	public class Startup
 	{
-		public Startup(IHostingEnvironment env)
+		public Startup(IWebHostEnvironment env)
 		{
 			// Setup configuration sources.
 			var builder = new ConfigurationBuilder()
@@ -81,10 +76,13 @@ namespace PopForums.Sample
 		}
 
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory)
 		{
 			// Records exceptions and info to the POP Forums database.
 			loggerFactory.AddPopForumsLogger(app);
+
+			// Enables languages
+			app.UsePopForumsCultures();
 
 			app.UseStaticFiles();
 
@@ -94,25 +92,28 @@ namespace PopForums.Sample
 			// Populate the POP Forums identity in every request.
 			app.UsePopForumsAuth();
 
-			// Wires up the SignalR hubs for real-time updates.
-			app.UsePopForumsSignalR();
-
 			app.UseDeveloperExceptionPage();
-
-			// Add MVC to the request pipeline.
-			app.UseMvc(routes =>
+			
+			// Add MVC to the request pipeline. The order of the next three lines matters:
+			app.UseRouting();
+			app.UseAuthorization();
+			app.UseEndpoints(endpoints =>
 			{
 				// POP Forums routes
-				routes.AddPopForumsRoutes(app);
+				endpoints.AddPopForumsEndpoints(app);
+
+				// need this if you have lots of routing and/or areas
+				endpoints.MapAreaControllerRoute(
+					"forumroutes", "forums",
+					"Forums/{controller=Home}/{action=Index}/{id?}");
 
 				// app routes
-
-				routes.MapRoute(
-					name: "areaRoute",
-					template: "{area:exists}/{controller=Home}/{action=Index}/{id?}");
-				routes.MapRoute(
-					name: "default",
-					template: "{controller=Home}/{action=Index}/{id?}");
+				endpoints.MapControllerRoute(
+					"areaRoute",
+					"{area:exists}/{controller=Home}/{action=Index}/{id?}");
+				endpoints.MapControllerRoute(
+					"default",
+					"{controller=Home}/{action=Index}/{id?}");
 			});
 		}
 	}
